reorganize to use only for python 2
