+++ tox.ini
@@ -159,7 +159,7 @@ isolated_build = true
 #      meaning running on py3.x will fail
 #      https://stackoverflow.com/questions/57459123/why-do-i-need-to-run-tox-twice-to-test-a-python-package-with-c-extension
 whitelist_externals=rm
-commands_pre=rm -f ddtrace/profiling/_build.c ddtrace/profiling/collector/stack.c ddtrace/profiling/collector/_traceback.c
+commands_pre=rm -f ddtrace/profiling/_build.c ddtrace/profiling/collector/stack.c ddtrace/profiling/collector/_traceback.c ddtrace/internal/_rand.c
              {envpython} {toxinidir}/setup.py develop
 usedevelop =
   # do not use develop mode with celery as running multiple python versions within
